Fix bart salmon on GH pages
When we do client-side routing the URL changes, but if you refresh or go directly to the page, you get a 404 page because it's not the
index.html which actually houses the app. GH pages does allow for a 404.html page. So we redirect from that page back to the index.html, but
save what URL we were on so that on the index.html we can manually replace the URL using the history API.

See: http://www.backalleycoder.com/2016/05/13/sghpa-the-single-page-app-hack-for-github-pages/