Migrate to Typescript + React hooks (#31)
Stripping out React Native support in order to migrate to TypeScript + React hooks. Also used this opportunity to move to Bart's (lousy) JSON API. The static data generated is now generated at build time so that it can remain fresh and won't get out of date like it did last time when new stations and routes were added.

Updated some of the salmon algorithm to handle the varying types of trains that happen during the days. Updated all of the scripts to be in TypeScript and to call the Bart API sequentially so it's not flooding it with requests.

The UI is the same and will be replaced in a following PR.

Lastly, will be moving to Netlify for hosting (with a redirect from benmvp.com) instead of hosting on GH Pages. That way everything works on the root domain.