Use new GA tracking code
Also removing the routing redirect logic (`404.html` page) since there is no routing and we've now switched to Netlify which likely has a different way to handle this problem.

Needed to switch to making scripts run in `http:` because the extra set up in Netlify seemed to cause the scripts to be flaky.