Add refresh support
Added new icon button to refresh the list of salmon routes / arrivals without having to refresh the whole page (which redownloads the app).

In a network-limited environment (like being in a station), this should hopefully provide a better user experience.

Also added a "last updated" message next to the refresh button to indicate when last the results had been updated.

Still holding off on auto-refresh because a rider may be continue to look at the results after boarding a train, and they wouldn't want the results to update which would cause them to lose the route they were following. We'd have to have some sort of route pinning/following functionality which updates as the salmon route progresses. Then we can also add auto-refresh.