Add refresh support (#39)
Added a new icon button to refresh the list of salmon routes/arrivals without having to refresh the whole page (which redownloads the app).

In a network-limited environment (like being in a station), this should hopefully provide a better user experience.

I also added a "last updated" message next to the refresh button to indicate when last the results had been updated.

Still holding off on auto-refresh because a rider may continue to look at the results after boarding a train, and they wouldn't want the results to update which would cause them to lose the route they were following. We'd have to have some sort of route pinning/following functionality which updates as the salmon route progresses. Then we can also add auto-refresh.

<img width="475" alt="Screen Shot 2019-11-05 at 8 25 01 PM" src="https://user-images.githubusercontent.com/5714478/68268277-64db7b80-000a-11ea-9d23-84ecf8aac690.png">