Auto generate data and store in git again (#42)
## Problem

The generation of the static data needed for the salmon algorithm is pretty flaky. The main offender is: `gen:station-routes`. Having to run it before tests and before build & deploy makes those jobs flaky.

In the beginning, the JSON files were stored in git because they became woefully out of date, especially when new stations like Antioch were added. That's why I switched to having it generated on-demand and ignored in git.

I want the data to be fresh, but it doesn't change frequently enough to be needed to be run _everytime_. We just need it to be periodically updated.

## Solution

So the solution is to leverage Github workflows & actions in order to run a cron job every night to run the data generation script. If any data changes, we'll push the new changes to `master`. As a result, running `npm test` & `npm build` no longer need to generate the data!. Running CI & Netflify should be **much** faster now.

Now that I somewhat know how Github workflows & actions work, I should be able to convert the `.travis.yml` over to a Github action as well in a follow-up PR!