Use Github workflow for CI (#44)
Now that we used a Github workflow for #43 we can transfer from Travis to Github. It leverages [`coverallsapp/github-action`](https://github.com/coverallsapp/github-action) to send the coverage data to [Coveralls](https://coveralls.io/). Also added a call to `npm run build` to verify that the build can be successfully generated.