Display "no routes" empty state (#46)
## Problem

Late at night or whenever there are no trains, there are no destinations. Therefore there are no salmon routes. We needed to fail gracefully by showing something and not just a blank screen.

In addition, when switching stations, we weren't clearing out the current salmon routes and arriving trains so there was a mismatch of data.


## Solution

Created an empty state component (`NoAvailableTrains`) to display when there are no arrivals. It uses a new sleeping emoji svg I found on the internet.

For the other problem, in the redux reducers when we change stations, I null out the salmon routes and arrivals. So `null` is a new supported and default state from empty.

Fixes #8.


## Screenshots

<img width="388" alt="Screen Shot 2019-11-26 at 4 50 51 PM" src="https://user-images.githubusercontent.com/5714478/69684661-5cf18300-106e-11ea-9c33-61acc4128f73.png">