Run unit tests before pushing new auto-gen data (#50)
## Problem

The pre-processed data changed substantially in [this commit](https://github.com/benmvp/bart-salmon/commit/58e4f77f74b2a51a4b0d6cff6a960273ff574716) from the nightly Github action cron. `routes.json` changed very dramatically. As a result, the tests running on mock data failed.


## Solution

We need to push a PR that can go through all of the checks to make sure we don't make `master` fail tests. So the solution, for now, is to ensure that the tests are run _before_ pushing the new data. If the tests fail, the cron will fail and we can run the job manually w/ a PR.