Add Settings screen (#51)
## Problem

There are a few configuration options that the users haven't been able to change:

- `riskinessFactor`
- `numArrivals`
- `numSalmonRoutes`

We need a screen/page that allows the user to change these AND the settings need to persist.

## Solution

Using `@reach/router`, added a new `/settings` route that displays a screen supporting changes these three configuration options.

Created a new `<Layout>` component that contains the bottom navigation where the user can switch between screens. Also rendering the `<Footer>` there so it appears underneath the page content for each screen. I needed to change up the look and feel of the footer so that the divider was inset.

Made "Safe" the default option for riskiness factor now that it can be configured.

Fixes #11 .


## Screenshots

Home page:

| Before | After |
| ------- | ----- |
| ![Screen Shot 2019-12-08 at 10 38 36 PM](https://user-images.githubusercontent.com/5714478/70413147-89f94a80-1a0b-11ea-8c36-40458f599ae4.png) | ![Screen Shot 2019-12-08 at 10 30 16 PM](https://user-images.githubusercontent.com/5714478/70413173-98476680-1a0b-11ea-997f-75881810d3f8.png) |

New settings screen:

![Screen Shot 2019-12-08 at 10 30 32 PM](https://user-images.githubusercontent.com/5714478/70413194-a4332880-1a0b-11ea-989e-d7cc43faee2d.png)