Create PR with updated meta (#56)
## Problem

The `auto-generate-data.yml` Github action creates new data, which can cause tests to fail ([for example](https://github.com/benmvp/bart-salmon/commit/8fe84e0a149c8d411b61dc623df7d88f88d080af/checks?check_suite_id=350581295)). With #50, we run the tests before pushing the new changes, but then that means that the script will continue to fail until we run `npm run gen:all` locally and then push the changes.


## Solution

The fix is to instead have the action, create a PR with the changes so that we can review it and then merge in the changes. This uses the [`peter-evans/create-pull-request`](https://github.com/marketplace/actions/create-pull-request) Github action. And if we take a long time to review the changes, new updates will continue to be pushed to the same branch (`auto-generated-data`), so the PR should always be up to date. Then when we land the PR and delete the branch, a new PR will get created pointing to the same branch.

Test PR: https://github.com/benmvp/bart-salmon/pull/58

Also:
- In `ci.yml` sending coverage data to `coveralls` _before_ verifying production build so we can get the results from them sooner. This essentially makes the production build verification and coveralls calculation parallel
- Minor styling change to `<Settings />` header